+++ packages/geo/src/index.ts
@@ -2,4 +2,5 @@ export * from "./color-scales";
 export * from "./camera-utils";
 export * from "./geography-cameras";
 export * from "./discovery-flyover";
+export * from "./cinematic-3d-cameras";
 export * from "./us-map";